Make **calc** it own seperate package
