+++ calc/__init__.py
@@ -1,0 +1,3 @@
+from .app import repl_executor, script_executor
+from .executor import calc_execute
+from .metadata import APPNAME, VERSION
