+++ calc/filter.py
