+++ calc/metadata.py
@@ -1,0 +1,2 @@
+APPNAME = "Calc"
+VERSION = "0.1.0"
