+++ calc/tokens.py
@@ -1,6 +1,6 @@
+from dataclasses import dataclass
 from enum import Enum
 from typing import Any
-from dataclasses import dataclass
 
 
 class TokenType(Enum):
