+++ calc/utils.py
@@ -1,22 +1,24 @@
-APPNAME = "Calc"
-VERSION = "0.1.0"
+from . import metadata
+
 
 def print_avaliable_commands():
     print("Available commands: ")
     print(" - quit \n - exit \n - help")
 
+
 def print_version():
-    print(f"{APPNAME} v{VERSION}")
+    print(f"{metadata.APPNAME} v{metadata.VERSION}")
+
 
 def print_help_msg(context = ""):
     if context == "quit" or context == "exit":
-        print(f"Command \"quit\" and \"exit\" are use to exit {APPNAME}!!")
+        print(f"Command \"quit\" and \"exit\" are use to exit {metadata.APPNAME}!!")
     elif context == "version":
         print_version()
     elif context == "command":
         print_avaliable_commands()
     else:
-        print(f"{APPNAME} | v{VERSION}")
+        print(f"{metadata.APPNAME} | v{metadata.VERSION}")
         print()
         print("Useage: python main.py <script>.calc")
         print(" OR ")
