+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "Calc"
+name = "calc"
 version = "0.1.0"
 description = "a simple python calculator scripting language"
 authors = ["Jeel Dobariya <dobariyaj34@gmail.com>"]
